refactor: 집계관련 타입, API 모듈 통합
카테고리 집계, 과소비 지수, 최대 소비 카테고리 정보에 대한 API, type을
aggregate 도메인으로 통합
