--- client/src/libs/api/aggregateCategory.ts
@@ -1,14 +1,0 @@
-import axios from '@/libs/axios';
-import endpoints from '@/libs/endpoints';
-import { AggregateData } from '@/commons/types/aggregateCategory';
-
-const aggregateCategoryAPI = {
-  getAggregateCategory: async (startDate: string, endDate: string): Promise<AggregateData> => {
-    const aggregateCategory = await axios.get<AggregateData>(endpoints.AGGREGATE_CATEGORY_API, {
-      params: { start: startDate, end: endDate },
-    });
-    return aggregateCategory;
-  },
-};
-
-export default aggregateCategoryAPI;
