+++ client/src/modules/aggregateCategory/actions.ts
@@ -1,6 +1,6 @@
 import { createAsyncAction } from 'typesafe-actions';
 import { AxiosError } from 'axios';
-import { AggregateData } from '@/commons/types/aggregateCategory';
+import { AggregateCategoryData } from '@/commons/types/aggregate';
 
 export const GET_AGGREGATE_CATEGORY = 'aggregateCategory/GET_AGGREGATE_CATEGORY';
 export const GET_AGGREGATE_CATEGORY_SUCCESS = 'aggregateCategory/GET_AGGREGATE_CATEGORY_SUCCESS';
@@ -10,4 +10,4 @@ export const getAggregateCategoryAsync = createAsyncAction(
   GET_AGGREGATE_CATEGORY,
   GET_AGGREGATE_CATEGORY_SUCCESS,
   GET_AGGREGATE_CATEGORY_FAILURE,
-)<string, AggregateData, AxiosError>();
+)<string, AggregateCategoryData, AxiosError>();
