+++ client/src/modules/aggregateCategory/thunks.ts
@@ -1,11 +1,11 @@
 /* eslint-disable import/prefer-default-export */
-import aggregateCategoryAPI from '@libs/api/aggregateCategory';
+import aggregateAPI from '@libs/api/aggregate';
 import createAsyncThunk from '@libs/createAsyncThunk';
 import { getAggregateCategoryAsync } from './actions';
 
 const getAggregateCategoryThunk = createAsyncThunk(
   getAggregateCategoryAsync,
-  aggregateCategoryAPI.getAggregateCategory,
+  aggregateAPI.getAggregateCategory,
 );
 
 export { getAggregateCategoryThunk };
