+++ client/src/modules/aggregateCategory/types.ts
@@ -1,4 +1,4 @@
-import { AggregateData } from '@/commons/types/aggregateCategory';
+import { AggregateCategoryData } from '@/commons/types/aggregate';
 import { ActionType } from 'typesafe-actions';
 import * as actions from './actions';
 
@@ -7,5 +7,5 @@ export type AggregateCategoryAction = ActionType<typeof actions>;
 export type AggregateCategoryState = {
   loading: boolean;
   error: Error | null;
-  data: AggregateData;
+  data: AggregateCategoryData;
 };
